+++ src/processor/linter/rules/index.ts
@@ -1,0 +1,2 @@
+export { BaseLintRule } from "./base";
+export { type TokenTypeValidator, TypeBasedRule } from "./TypeBasedRule";